+++ src/RCarousel/__tests__/tools/render.ts
@@ -1,0 +1,15 @@
+import { render as tlRender } from '@testing-library/react';
+import { finishAnim } from './springConfig';
+
+export const render = ((...args: any[]) => {
+  const renderResult = finishAnim(() =>
+    // @ts-ignore
+    tlRender(...args)
+  );
+  const originalRerender = renderResult.rerender;
+
+  renderResult.rerender = (...args: any[]) =>
+    // @ts-ignore
+    finishAnim(() => originalRerender(...args));
+  return renderResult;
+}) as typeof tlRender;