+++ src/RCarousel/helpers/types.ts
@@ -1,0 +1,4 @@
+export interface DisplayRule {
+  breakpoint?: number;
+  value: number;
+}