+++ src/hooks/__mocks__/useOnResize.ts
@@ -1,12 +1,12 @@
 import React from 'react';
 
 export function useOnResize(containerRef: React.RefObject<HTMLDivElement>) {
-  const [size, setSize] = React.useState([0, 0]);
+  const [size, setSize] = React.useState({ width: 0, height: 0 });
   React.useEffect(() => {
     if (!containerRef.current) {
       return;
     }
-    setSize([window.innerWidth, window.innerHeight]);
+    setSize({ width: window.innerWidth, height: window.innerHeight });
   }, [containerRef]);
   return size;
 }