+++ stories/RCarousel/btns/index.ts
@@ -1,0 +1,4 @@
+import ExampleBackBtn from './ExampleBackBtn';
+import ExampleNextBtn from './ExampleNextBtn';
+
+export { ExampleBackBtn, ExampleNextBtn };