+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+# [1.1.0](https://github.com/demidyuk/recycling-react-carousel/compare/v1.0.2...v1.1.0) (2021-04-25)
+
+
+### Bug Fixes
+
+* change error conditions for maxItemSize and displayAtOnce ([d15260a](https://github.com/demidyuk/recycling-react-carousel/commit/d15260a3caa3c7d676c627f9a408290f04858dd8))
+
+
+### Features
+
+* add additional parameter in onRangeChange for count of slides ([4a48775](https://github.com/demidyuk/recycling-react-carousel/commit/4a487756d34c234e078e154c46220cd82723663e))
+* add change reasons and optimize updates ([242c01f](https://github.com/demidyuk/recycling-react-carousel/commit/242c01fb78ba74aeaad69ffc583397bfeccc93f8))
+* add cursor shifts when adding new slides ([209b288](https://github.com/demidyuk/recycling-react-carousel/commit/209b28892d824cc78d92420fd51b97fbb3068b13))
+* add defaultCursor prop and snapshot testing ([6bc677e](https://github.com/demidyuk/recycling-react-carousel/commit/6bc677e63f1f9afe07b4cbf9a725a97e09b56dc8))
+* add loop mode ([ec59182](https://github.com/demidyuk/recycling-react-carousel/commit/ec59182ab1e151e7966c42ba40ed3cfb84c22d08))
+* add swipeThreshold prop ([b0390c7](https://github.com/demidyuk/recycling-react-carousel/commit/b0390c70ef17c3075872e7ca141f8adf656ba239))
+* add trimEnd prop ([be917dd](https://github.com/demidyuk/recycling-react-carousel/commit/be917dd812cb9d400b6c849fe703635e23093fef))
+* add userSelect and springConfig props, enhance displayAtOnce to support breakpoints ([231312e](https://github.com/demidyuk/recycling-react-carousel/commit/231312e375afd8693b5ee44921c49b3d709a7643))
+* refactor code in useCursor hook and add  goTo method ([39ae9d2](https://github.com/demidyuk/recycling-react-carousel/commit/39ae9d28217ee6144b636c18bf405a62e38caa24))
+
+## [1.0.2](https://github.com/demidyuk/recycling-react-carousel/compare/v1.0.1...v1.0.2) (2021-03-18)
+
+
+### Bug Fixes
+
+* fix bug with init prop of useCursor ([6fa2782](https://github.com/demidyuk/recycling-react-carousel/commit/6fa2782f0ef28cfdbcf7da6cc81b67db67ffeaab))
+
+## [1.0.1](https://github.com/demidyuk/recycling-react-carousel/compare/v1.0.0...v1.0.1) (2021-03-14)
+
+
+### Bug Fixes
+
+* make item wrapper width 100% on the y-axis ([82ae8b9](https://github.com/demidyuk/recycling-react-carousel/commit/82ae8b915609aef1c912e374d9a3befa37c1e18a))
+
 # 1.0.0 (2021-03-13)